pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airflow-file-auth-manager"
version = "0.1.0"
description = "YAML file-based lightweight Auth Manager for Apache Airflow 3.x"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Yeonguk", email = "yeonguk@example.com" }
]
keywords = ["airflow", "authentication", "auth-manager"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Apache Airflow",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "apache-airflow>=3.0.0",
    "bcrypt>=4.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yeonguk/airflow-file-auth-manager"
Repository = "https://github.com/yeonguk/airflow-file-auth-manager"

[project.entry-points."airflow.auth.manager"]
file_auth_manager = "airflow_file_auth_manager:FileAuthManager"

[project.entry-points."airflow.plugins"]
file_auth_cli = "airflow_file_auth_manager.cli:FileAuthCLIPlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_file_auth_manager"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/examples",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
